+++ packages/interweave/src/types.js
@@ -33,8 +33,8 @@ export type AfterParseCallback = (content: React$Node[], props: Object) => React
 export type BeforeParseCallback = (content: string, props: Object) => string;
 
 export interface FilterInterface {
-  attribute: string,
-  filter(value: string): string,
+  attribute(name: string, value: string): string,
+  node(name: string, node: NodeInterface): NodeInterface,
 }
 
 export interface MatcherInterface {